pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leadfinder"
version = "0.1.0"
description = "LeadFinder: discover and enrich business leads by query and geography"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
]

[tool.setuptools.packages.find]
include = ["leadfinder*"]
exclude = ["data*", "notebooks*"]

[tool.ruff]
line-length = 100
